requirements.txt
streamlit==1.32.0
pandas==2.2.0
torch
tweepy>=4.14.0
wordcloud>=1.9.3
praw==7.7.1
newsapi-python==0.2.7
python-dotenv==1.0.0
transformers>=4.37.0
nltk>=3.8.1